move registration commands under deviceregistration directly
