+++ api/spec/json/deviceCredentials.json
@@ -85,8 +85,8 @@
     {
       "name": "registerNewDevice",
       "method": "POST",
-      "description": "Register device",
-      "descriptionLong": "Register a new device (request)",
+      "description": "Register device with username/password and manual device approval/bootstrapping",
+      "descriptionLong": "Register a new device (request) where the device is using the manual device bootstrapping\nprocess to retrieve its device credentials (username/password).\n\nSee Cumulocity docs for more details: https://cumulocity.com/docs/2024/device-integration/rest/\n",
       "path": "devicecontrol/newDeviceRequests",
       "accept": "application/vnd.com.nsn.cumulocity.newDeviceRequest+json",
       "alias": {
