+++ api/spec/yaml/deviceCredentials.yaml
@@ -64,8 +64,12 @@ commands:
 
   - name: registerNewDevice
     method: POST
-    description: Register device
-    descriptionLong: Register a new device (request)
+    description: Register device with username/password and manual device approval/bootstrapping
+    descriptionLong: |
+      Register a new device (request) where the device is using the manual device bootstrapping
+      process to retrieve its device credentials (username/password).
+
+      See Cumulocity docs for more details: https://cumulocity.com/docs/2024/device-integration/rest/
     path: devicecontrol/newDeviceRequests
     accept: application/vnd.com.nsn.cumulocity.newDeviceRequest+json
     alias:
