+++ pkg/cmd/deviceregistration/register/register.auto.go
@@ -24,15 +24,19 @@ type RegisterCmd struct {
 	factory *cmdutil.Factory
 }
 
-// NewRegisterCmd creates a command to Register device
+// NewRegisterCmd creates a command to Register device with username/password and manual device approval/bootstrapping
 func NewRegisterCmd(f *cmdutil.Factory) *RegisterCmd {
 	ccmd := &RegisterCmd{
 		factory: f,
 	}
 	cmd := &cobra.Command{
 		Use:   "register",
-		Short: "Register device",
-		Long:  `Register a new device (request)`,
+		Short: "Register device with username/password and manual device approval/bootstrapping",
+		Long: `Register a new device (request) where the device is using the manual device bootstrapping
+process to retrieve its device credentials (username/password).
+
+See Cumulocity docs for more details: https://cumulocity.com/docs/2024/device-integration/rest/
+`,
 		Example: heredoc.Doc(`
 $ c8y deviceregistration register --id "ASDF098SD1J10912UD92JDLCNCU8"
 Register a new device
