+++ pkg/cmd/root/root.go
@@ -43,7 +43,7 @@ import (
 	deviceManagementCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/devicemanagement"
 	deviceprofilesCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/deviceprofiles"
 	deviceregistrationCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/deviceregistration"
-	"github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/deviceregistration/bulk"
+	deviceregistrationBulkCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/deviceregistration/registerBulk"
 	devicesCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/devices"
 	devicesAssertCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/devices/assert"
 	devicesAvailabilityCmd "github.com/reubenmiller/go-c8y-cli/v2/pkg/cmd/devices/availability"
@@ -408,8 +408,9 @@ func NewCmdRoot(f *cmdutil.Factory, version, buildDate string) *CmdRoot {
 	cmd.AddCommand(agents)
 
 	deviceregistration := deviceregistrationCmd.NewSubCommand(f).GetCommand()
-	bulkDeviceRegistration := bulk.NewSubCommand(f).GetCommand()
-	deviceregistration.AddCommand(bulkDeviceRegistration)
+	deviceregistration.AddCommand(deviceregistrationBulkCmd.NewRegisterBasicCmd(f).GetCommand())
+	deviceregistration.AddCommand(deviceregistrationBulkCmd.NewRegisterCumulocityCACmd(f).GetCommand())
+	deviceregistration.AddCommand(deviceregistrationBulkCmd.NewRegisterExternalCACmd(f).GetCommand())
 	cmd.AddCommand(deviceregistration)
 
 	// microservices
